Begin adding a layman's-terms description of Endpoints' functionality
The existing What Is It? section did not fully explain to me how Endpoints was
intended to be used, so I've taken a stab at articulating in a little more detail
the purpose of the library and its relationship to your existing tools. I find
these sorts of overviews to be personally useful, and would love input on how to
improve the description so that it can help expand Endpoints' potential audience.

----

Consolidate short blurb and unify "what is it" and "layman's" headers

The repo description and existing "What Is It" text overlapped in purpose,
and the "Layman's Terms" heading isn't exactly inviting, so I flipped the
existing "what is it" text up to be an abstract of sorts for the README
and consolidated the new descriptive text and example repo link under the
"What is it" heading.

---

Clean up wording per PR feedback

closes gh-87